+++ steam-cleaner/src/main/java/org/gametools/cleaner/actions/GetApp.java
@@ -12,11 +12,11 @@ public class GetApp implements ActionRunner {
 
     private final StorageLocator storageLocator;
     private final Function<StorageDrive, AppsRepository> appsRepositoryFactory;
-    private final Integer instanceId;
+    private final Long instanceId;
 
     public GetApp(StorageLocator storageLocator,
                   Function<StorageDrive, AppsRepository> appsRepositoryFactory,
-                  Integer instanceId) {
+                  Long instanceId) {
         this.storageLocator = storageLocator;
         this.appsRepositoryFactory = appsRepositoryFactory;
         this.instanceId = instanceId;
